+++ format_bad_2.py
@@ -1,3 +1,7 @@
 """"""
 
 __all__ = []
+import json
+
+
+my_thing = {}