+++ format_good_2.py
@@ -1,3 +1,8 @@
 """"""
 
 __all__ = []
+
+import json
+
+
+my_thing = {}