+++ tests/unit/test_mock.py
@@ -1,3 +1,4 @@
+"""TODO: Add module docstring."""
 
 
 def test_thing():